Fix initialisation of supplementary groups
Commit e9e2e369 initialised the supplementary groups of the new
user. However, the call to initgroups() came after the
setuid/setgid. initgroups() is based on the setgroups() system
call which on Linux needs the CAP_SETGID capability.

This worked in my simple testing as I was testing running 'id'
as root. This however fails for lesser users.

Move the initgroups() call to before the setuid().

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>