Enhance logging
We log PWD which is the directory that the command being run from
runasuser was run from, which by default will be the home directory
of the user that the command is being run from runasuser as.

So we now log EWD, which is the directory that the _runasuser_ command
was run from.

We also now log the arguments of the command being run from runasuser.

Here's an example of the new log output:

Aug 22 19:02:08 thor runasuser: andrew : TTY=pts/3 ; EWD=/tmp ;
PWD=/home/nobody ; USER=nobody ; COMMAND=/bin/ls -l data/

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>