Fix a segfault when running from cron
When we log to syslog, one of the things we log is the tty. However when
running from cron, there is no tty and this comes back as NULL, which
then causes a segfault in syslog()

Fix this by checking for a NULL tty and if so use '(none)' as the tty
name.

i.e

    TTY=(none)

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>