Re-work strtok() usage
The main problem it had was leaking memory, by making a copy of a string
to pass to strtok then subsequently setting it to NULL, then trying to
free it when finished. free(NULL) is a noop.

Re-work it so we don't need to set string to NULL, we also have a nice
code reduction and simpler code after this.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>