Fix broken command logging
Commit 94e9674 ("Fix a compiler warning") broke the logging of commands
(COMMAND=)

When running something like

    ./runasuser nobody ls -l

we should get the following logged for COMMAND

    COMMAND=/bin/ls -l

however due to the added snprintf() call not taking into account what
was already written into the buffer before adding in the command args
we'd end up with

    COMMAND= -l

Fix is simple. Just track how much data we write into the buffer and use
that as an offset into the buffer for each write.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>